Keep regional macOS audit identities with upstream version overrides. Adopt maintained runtime sources and update regional tests for source and save-path migrations. Retire the obsolete WFC patch; trace activation follows upstream runtime behavior.

+++ runtime/src/android/controller_mapping.cpp
@@ -5,15 +5,15 @@
 namespace kartpad::android {
 namespace {
 
-constexpr uint32_t Pack(const ControllerButtonMapping& mapping) noexcept {
-  uint32_t packed = 0;
+constexpr uint64_t Pack(const ControllerButtonMapping& mapping) noexcept {
+  uint64_t packed = 0;
   for (std::size_t index = 0; index < mapping.size(); ++index) {
-    packed |= static_cast<uint32_t>(mapping[index]) << (index * 4);
+    packed |= static_cast<uint64_t>(mapping[index]) << (index * 4);
   }
   return packed;
 }
 
-std::atomic<uint32_t> g_mapping{Pack(kDefaultControllerButtonMapping)};
+std::atomic<uint64_t> g_mapping{Pack(kDefaultControllerButtonMapping)};
 
 }  // namespace
 
@@ -26,7 +26,7 @@ void PublishControllerButtonMapping(
 
 ControllerButtonMapping ReadControllerButtonMapping() noexcept {
   ControllerButtonMapping mapping{};
-  const uint32_t packed = g_mapping.load(std::memory_order_acquire);
+  const uint64_t packed = g_mapping.load(std::memory_order_acquire);
   for (std::size_t index = 0; index < mapping.size(); ++index) {
     mapping[index] = static_cast<uint8_t>((packed >> (index * 4)) & 0x0f);
   }
